+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -6,7 +6,7 @@ object GradlePlugins {
 
     private object Versions {
         const val dokka = "1.6.10"
-        const val kotlin = "1.5.31"
+        const val kotlin = "1.6.10"
         const val versions = "0.42.0"
     }
 
@@ -20,7 +20,7 @@ object Libs {
     private object Versions {
         const val junitJupiter = "5.8.2"
         const val kotlinCoroutines = "1.6.0"
-        const val moshi = "1.12.0"
+        const val moshi = "1.13.0"
         const val okhttp = "3.12.13"
         const val retrofit = "2.6.4"
         const val threetenbp = "1.6.0"