+++ engelsystem-base/src/test/kotlin/info/metadude/kotlin/library/engelsystem/ProductionApiTest.kt
@@ -10,7 +10,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import org.junit.jupiter.api.Assertions.fail
 import org.junit.jupiter.api.Test
 import org.threeten.bp.ZonedDateTime
-import retrofit2.awaitResponse
+import retrofit2.HttpException
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class ProductionApiTest {
@@ -26,14 +26,11 @@ class ProductionApiTest {
     @Test
     fun `Validates a successful shifts response`() = runTest {
         try {
-            val response = service.getShifts(URL_PART_PATH, VALID_API_KEY).awaitResponse()
-            if (response.isSuccessful) {
-                val shifts = response.body()
-                assertThat(shifts).isNotNull()
-                shifts?.forEach {
-                    assertThat(it).isNotNull()
-                    assertShift(it)
-                }
+            val shifts = service.getShifts(URL_PART_PATH, VALID_API_KEY)
+            assertThat(shifts).isNotNull()
+            shifts.forEach {
+                assertThat(it).isNotNull()
+                assertShift(it)
             }
         } catch (t: Throwable) {
             fail("Should not throw $t")
@@ -63,13 +60,13 @@ class ProductionApiTest {
     @Test
     fun `Validates a failure shifts response`() = runTest {
         try {
-            val response = service.getShifts(URL_PART_PATH, INVALID_API_KEY).awaitResponse()
-            assertThat(response.isSuccessful).isFalse()
-            assertThat(response.code()).isEqualTo(403)
-            assertThat(response.body()).isNull()
-            assertThat(response.errorBody()).isNotNull()
-        } catch (t: Throwable) {
-            fail("Should not throw $t")
+            service.getShifts(URL_PART_PATH, INVALID_API_KEY)
+            fail("Request should not succeed.")
+        } catch (e: HttpException) {
+            assertThat(e.message()).isEqualTo("Forbidden")
+            assertThat(e.code()).isEqualTo(403)
+            assertThat(e.response()!!.body()).isNull()
+            assertThat(e.response()!!.errorBody()).isNotNull()
         }
     }